show instructors and students on people page (#25)
