+++ Components/Sidebar.tsx
@@ -9,7 +9,6 @@ import OpenGraderLogo from "./OpenGraderLogo";
 import { useAppDispatch, useAppSelector } from "hooks";
 import { loadUser, setRole } from "store/userSlice";
 
-
 const classNames = (...classes: string[]): string => classes.filter(Boolean).join(" ");
 
 const SideBarLinks = () => {
@@ -18,7 +17,7 @@ const SideBarLinks = () => {
   const routes = [
     { text: "Courses", link: "", icon: HiOutlineAcademicCap },
     { text: "Assignments", link: ":course/assignment", icon: HiOutlineClipboard },
-    { text: "Students", link: ":course/student", icon: HiOutlineUsers },
+    { text: "People", link: ":course/people", icon: HiOutlineUsers },
     { text: "Reports", link: ":course/report", icon: HiOutlineChartBar },
   ];
